Add a dispatch.yaml configuration for serving in woolaroo dir
This commit updates the server configuration script and cloud build
to set GAE to serve the contents at a subdirector level.

This is because this is how we serve it at Google, and isn't a
requirement. This means we can configure a staging environment
to be identical to prod and catch any issues early.

Relates to #179